Extract `FraudProofGenerator` into a seperate `fraud_proof` module
The fraud proof can be constructed in two places:

- When the executor verifies the receipt received from the executor gossip network.
- When the executor checks the receipts included in each primary block.

This struct will be used by both `Executor` and `BundleProcessor` later.

No logical changes in this commit, the changes are basically to make it
ocmpile.
